Imported Upstream version 2.11

+++ ykpamcfg.c
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2011 Yubico AB.
+ * Copyright (c) 2011-2012 Yubico AB
  * All rights reserved.
  *
  * Author : Fredrik Thulin <fredrik@yubico.com>
@@ -155,7 +155,7 @@ do_add_hmac_chalresp(YK_KEY *yk, uint8_t slot, bool verbose, char *output_dir, i
     goto out;
 
   if (response_len > sizeof (state.response)) {
-    fprintf (stderr, "Got too long response ??? (%i/%i)", response_len, sizeof(state.response));
+    fprintf (stderr, "Got too long response ??? (%u/%lu)", response_len, (unsigned long) sizeof(state.response));
     goto out;
   }
   memcpy (state.response, buf, response_len);
